gemma: own paged runtime directly
